bsf: fix filtering for several network address/mask
This patch fixes kernel-space filtering via BSF for several
network addresses. The problem is that we store the network
address of the netlink message in the ALU. Then, we perform
an AND of the network mask and the address, this operation
is stored again in the ALU. If we compare the address with
a second address, we have to reload the address to the ALU.

The following example clarifies the problem, in the following
order, we want to filter:

1) 224.0.0.0/4
2) 127.0.0.1/32

Now, we receive traffic from 127.0.0.1, it should be filtered.
However, without this patch, it is not. Let's see why:

ALU   7f000001 (addr=127.0.0.1)
AND   f0000000 (cidr=4)
-------------------------------
ALU   70000000

this is stored in the ALU. Then, we check for 127.0.0.1:

ALU   70000000 (addr=127.0.0.1) <-- it should be 7f000001
AND   ffffffff (cidr=32)
-------------------------------
ALU   70000000

This does not match 7f000001. To fix this, we have to reload
7f000001 to the ALU. Thus, the second comparison works fine.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>