conntrack: labels: labelmap_new: make sure errno is 0 when no labels are found
nfct_labelmap_new returns NULL on failure, e.g. when file cannot be
opened.  It will also fail if no labels have been parsed, and in this
case, content of errno is random.

Avoid it by making sure that errno is re-set when no labels were found.

While at it, also change ptr test when parsing so reviewers don't
need to triple check that this cannot result in out-of-bounds read.

Reported-by: Afschin Hormozdiary <Afschin.Hormozdiary@sophos.com>
Signed-off-by: Florian Westphal <fw@strlen.de>